refactor(preprocess): move the preprocess logic into Preprocess.hs
